Pass user data as state
