+++ client/src/pages/ListPage.js
@@ -1,0 +1,4 @@
+export default function ListPage(inputValue) {
+  console.log(inputValue);
+  return <div> Hello Overview</div>;
+}
